draw method called when canvas is clicked
class is initialized on webpage startup and when the canvas is clicked, the draw function within the class is called. Currently it only outputs to the console. The file 'app.js' is now redundant. Once the code within has been relocated I will delete it. All relevant js code is in 'canvas.js'