canvas resizes and app.js deleted
the canvas now resized when height and width are input. This also sidesteps the visual bug caused by the rounding error as now the height and width pixels will always be evenly divisible. Also, the app.js file was redundant, but had not yet been deleted. It has now been removed.